add test for registration, ad test data, add list of complete urls and update requirements.txt with new requirements
