+++ authentication/tests/__init__.py
@@ -1,0 +1,2 @@
+from .BaseTestApi import BaseTestApi
+
