+++ staticfiles/test_data.py
@@ -1,11 +1,16 @@
-User_authentication = {
+User_authentication_1 = {
     'username': 'test_user_1',
     'email': 'test-user_1@mail.com',
     'password': 'password11'
 }
-User_api = {
+User_authentication_2 = {
     'username': 'test_user_2',
     'email': 'test-user_2@mail.com',
+    'password': 'qwerty1234567890`!@#$%^&*()_+=?>\';[]\\{}?|/'
+}
+User_api = {
+    'username': 'test_user_3',
+    'email': 'test-user_3@mail.com',
     'password': 'password11'
 }
 Board_api_create_new_1 = {
